backend/requirements.txt
google-generativeai
fastapi
uvicorn
python-multipart
fpdf
pillow

